refactor(renderer): extract launcher inline arguments
